Refactor: Use panel components
