+++ src/components/yeti/YetiPanel.vue
@@ -1,0 +1,78 @@
+<template>
+  <div class="yeti-panel">
+    <slot></slot>
+  </div>
+</template>
+
+<script>
+export default {};
+</script>
+
+<style scoped lang="scss">
+@import '@/assets/sass/variables.scss';
+
+/deep/ {
+  p:not(:last-child),
+  ul:not(:last-child),
+  table:not(:last-child) {
+    margin-bottom: 1rem;
+  }
+
+  a:hover,
+  a:focus {
+    text-decoration: underline;
+  }
+
+  h4 {
+    margin-top: 1rem;
+    margin-bottom: 0.25rem !important;
+  }
+
+  abbr {
+    text-decoration: none;
+    border-bottom: dotted 2px #aaa;
+  }
+
+  strong,
+  dt {
+    font-weight: bold;
+    color: $grey;
+  }
+
+  .content-ul {
+    list-style: disc outside;
+    margin-left: 2em;
+  }
+
+  .yetiform-info {
+    font-size: 0.8em;
+    opacity: 0.75;
+  }
+
+  .yetiform-note {
+    position: relative;
+    font-size: 0.9em;
+    padding: 1em;
+    padding-left: 4em;
+    margin-bottom: 2em;
+    background-color: $white-ter;
+    border-radius: 2px;
+  }
+
+  .yetiform-note::before {
+    content: '?';
+    position: absolute;
+    top: 1em;
+    left: 1em;
+    color: white;
+    font-size: 1.14em;
+    font-weight: bold;
+    text-align: center;
+    height: 1.375em;
+    width: 1.375em;
+    line-height: 1.4em;
+    background-color: rgba(0, 60, 136, 0.5);
+    border-radius: 2px;
+  }
+}
+</style>
